requirements.txt
openai==1.11.0
python-dotenv==1.0.1
